chore: moved retrieveRemarkFromTransaction function to an use case
