+++ backend/src/services/transactionManager/index.ts
@@ -1,5 +1,4 @@
 export { createApi, createKeyPair } from "../../utils/networkApi.js";
 export { getAccountNonce } from "./queue.js";
-export { retrieveRemarkFromTransaction } from "./retrieveTransactionData.js";
 export { createTransactionManager } from "./transactionManager.js";
 export { Transaction, TransactionResult } from "./types.js";
